splash page added, and finished mvc refactoring
