+++ app/src/main/java/com/example/afroaroma/view/AdminHomeActivity.kt
@@ -1,19 +1,14 @@
 package com.example.afroaroma.view
 
-import android.content.ContentValues
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.widget.Button
-import android.widget.ImageView
 import android.widget.TextView
 import com.example.afroaroma.R
-import com.example.afroaroma.admin.controller.AuthController
-import com.example.afroaroma.admin.controller.FirestoreController
-import com.google.firebase.auth.FirebaseAuth
+import com.example.afroaroma.controller.AuthController
+import com.example.afroaroma.controller.FirestoreController
 import com.google.firebase.auth.ktx.auth
-import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 
 class AdminHomeActivity : AppCompatActivity() {
