+++ app/src/main/res/anim/fade_out.xml
@@ -1,0 +1,7 @@
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:fromAlpha="1.0"
+        android:toAlpha="0.0"
+        android:duration="500"
+        android:interpolator="@android:interpolator/decelerate_cubic" />
+</set>
