+++ app/src/main/res/anim/slide_in_right.xml
@@ -1,0 +1,4 @@
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate android:fromXDelta="100%p" android:toXDelta="0"
+        android:duration="500"/>
+</set>
