+++ app/src/main/res/anim/slide_out_left.xml
@@ -1,0 +1,4 @@
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate android:fromXDelta="0" android:toXDelta="-100%p"
+        android:duration="500"/>
+</set>
